Change float to int...

+++ res/shaders/batch/frag.glsl
@@ -4,10 +4,10 @@ layout(location = 1) uniform sampler2D u_textures[8];
 
 in vec4 pass_color;
 in vec2 pass_uv;
-flat in float pass_tex;
+flat in int pass_tex;
 
 out vec4 color;
 
 void main() {
-    color = texture(u_textures[int(pass_tex)], pass_uv) * pass_color;
+    color = texture(u_textures[pass_tex], pass_uv) * pass_color;
 }
